refactor(lex-babel): split flat_to_nested and render_dispatch into submodules
